# docs/openapi-consumer.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "X-Ray 企业变化解释器",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/companies": {
      "get": {
        "summary": "Companies",
        "operationId": "companies_api_companies_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sources/{source_id}": {
      "get": {
        "summary": "Source",
        "operationId": "source_api_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sources/{source_id}/document": {
      "get": {
        "summary": "Local Document",
        "operationId": "local_document_api_sources__source_id__document_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/companies": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Companies",
        "operationId": "companies_api_v2_companies_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/companies/{company_id}/snapshot": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Snapshot",
        "operationId": "snapshot_api_v2_companies__company_id__snapshot_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/companies/{company_id}/timeline": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Timeline",
        "operationId": "timeline_api_v2_companies__company_id__timeline_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/companies/{company_id}/coverage": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Coverage",
        "operationId": "coverage_api_v2_companies__company_id__coverage_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/companies/{company_id}/outcomes": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Outcomes",
        "operationId": "outcomes_api_v2_companies__company_id__outcomes_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Id"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "As Of"
            }
          },
          {
            "name": "window_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 730,
              "minimum": 1,
              "default": 180,
              "title": "Window Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/sources/{document_id}": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Source",
        "operationId": "source_api_v2_sources__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/sources/{document_id}/document": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Document",
        "operationId": "document_api_v2_sources__document_id__document_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/sources/{document_id}/pages/{page}": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Page Image",
        "operationId": "page_image_api_v2_sources__document_id__pages__page__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/png": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/sources/{document_id}/viewer": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Viewer",
        "operationId": "viewer_api_v2_sources__document_id__viewer_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/comparisons": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Compare",
        "operationId": "compare_api_v2_comparisons_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComparisonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComparisonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/comparisons/{comparison_id}": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Comparison",
        "operationId": "comparison_api_v2_comparisons__comparison_id__get",
        "parameters": [
          {
            "name": "comparison_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Comparison Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/search": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Search",
        "operationId": "search_api_v2_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80,
              "title": "Query"
            }
          },
          {
            "name": "as_of",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "As Of"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/discovery": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Discovery",
        "description": "Network discovery is opt-in; no request writes verified company evidence.",
        "operationId": "discovery_api_v2_discovery_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 80,
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComparisonRequest": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "window_days": {
            "type": "integer",
            "maximum": 730.0,
            "minimum": 1.0,
            "title": "Window Days",
            "default": 180
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "as_of"
        ],
        "title": "ComparisonRequest"
      },
      "ComparisonResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "method_version": {
            "type": "string",
            "title": "Method Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "matches": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Matches"
          },
          "excluded": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Excluded"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "company_id",
          "as_of",
          "window_days",
          "method_version",
          "status",
          "matches",
          "excluded",
          "candidate_count",
          "dataset_version",
          "limitations"
        ],
        "title": "ComparisonResponse"
      },
...
      "EntityPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntityView"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "items",
          "total",
          "dataset_version"
        ],
        "title": "EntityPage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutcomeResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "title": "Company Id"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "window_days": {
            "type": "integer",
            "title": "Window Days"
          },
          "window_end": {
            "type": "string",
            "title": "Window End"
          },
          "observation_status": {
            "type": "string",
            "title": "Observation Status"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/OutcomeView"
            },
            "type": "array",
            "title": "Outcomes"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "company_id",
          "as_of",
          "window_days",
          "window_end",
          "observation_status",
          "outcomes",
          "dataset_version",
          "limitations"
        ],
        "title": "OutcomeResponse"
      },
...
      "SnapshotResponse": {
        "properties": {
          "company": {
            "$ref": "#/components/schemas/EntityView"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/FactView"
            },
            "type": "array",
            "title": "Facts"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineItem"
            },
            "type": "array",
            "title": "Timeline"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentIndex"
            },
            "type": "array",
            "title": "Documents"
          },
          "coverage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Coverage"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "company",
          "as_of",
          "dataset_version",
          "facts",
          "timeline",
          "documents",
          "coverage",
          "limitations"
        ],
        "title": "SnapshotResponse"
      },
      "SourceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "published_at": {
            "type": "string",
            "title": "Published At"
          },
          "fetched_at": {
            "type": "string",
            "title": "Fetched At"
          },
          "extraction_status": {
            "type": "string",
            "title": "Extraction Status"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "integrity_status": {
            "type": "string",
            "title": "Integrity Status"
          },
          "local_available": {
            "type": "boolean",
            "title": "Local Available"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "title",
          "url",
          "sha256",
          "published_at",
          "fetched_at",
          "extraction_status",
          "verification_status",
          "integrity_status",
          "local_available"
        ],
        "title": "SourceResponse"
      },
...
    }
  }
}